⬆ Go 1.19 (#158)
This is an upgrade to Go 1.19. Still the 1.18 tests are required for merging to main because there are services using this library that run on 1.18 atm.

https://github.com/golangci/golangci-lint-action/issues/545#issuecomment-1209931303

* Replace `io/ioutil` with `os` functions

Fixes Linter warning:
SA1019: "io/ioutil" has been deprecated since Go 1.16: As of Go 1.16, the same functionality is now provided by package io or package os, and those implementations should be preferred in new code. See the specific function documentation for details.  (staticcheck)